% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictSoftsplits}
\alias{predictSoftsplits}
\title{Prediction according to `soft tree'.}
\usage{
predictSoftsplits(fit, newdata)
}
\arguments{
\item{fit}{The soft tree.}

\item{newdata}{Data to classify.}
}
\value{
The matrix of predicted class probabilities.
}
\description{
Prediction according to `soft tree'.
}
