% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds_by_esd.R
\name{softening.by.esd}
\alias{softening.by.esd}
\title{Soften splits separately using error standard deviation.}
\usage{
softening.by.esd(fit, d)
}
\arguments{
\item{fit}{The soft tree}

\item{d}{The data set}
}
\description{
Set boundaries determined by given data to the splits in the tree,
such that in any inner node if its splitting value would be moved there,
then the number of misclassified cases in this node would be one standard deviation over the actual misclassification.
}
\details{
This is the same approach as C4.5 uses for "probabilistic splits"
}
\references{
{Quinlan, J. Ross} {(1993)}, \emph{C4.5: programs for machine learning}, {San Francisco, CA, USA}: {Morgan Kaufmann Publishers Inc.}
}
